Refactor loading of theme Pro Price helper
Prepending module to the Pro Price helper was causing issues in CI only.
We don't actually have to do this as helper modules are all included
into the same context via ActionView::Base.
