+++ lib/views/alaveteli_pro/plans/_pricing_tiers.html.erb
@@ -16,7 +16,7 @@
       </p>
 
       <p class="price-label-option">
-        <%= link_to short_alternative_price_text(@prices[1]), plan_path(@prices[1]) %>
+        <%= link_to alternative_price_text(@prices[1]), plan_path(@prices[1]) %>
       </p>
     </div>
